rename and move a file
