+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Client } from "boardgame.io/react";
 import "./App.css";
-import { Slashchain } from "./slashchain/view/game";
+import { Slashchain } from "./slashchain/game";
 import { GameComponent } from "./slashchain/view/game_state";
 
 const App = Client({
